+++ sdk/search/Azure.Search.Documents/src/Indexes/FieldBuilder.cs
@@ -222,9 +222,13 @@ namespace Azure.Search.Documents.Indexes
             static bool IsNullableType(Type type) =>
                 type.IsConstructedGenericType && type.GetGenericTypeDefinition() == typeof(Nullable<>);
 
-            if (s_primitiveTypeMap.TryGetValue(propertyType, out SearchFieldDataType SearchFieldDataType))
+            if (s_primitiveTypeMap.TryGetValue(propertyType, out SearchFieldDataType searchFieldDataType))
             {
-                return DataTypeInfo.Simple(SearchFieldDataType);
+                return DataTypeInfo.Simple(searchFieldDataType);
+            }
+            else if (GeometryFactory.CanCreate(propertyType))
+            {
+                return DataTypeInfo.Simple(SearchFieldDataType.GeographyPoint);
             }
             else if (IsNullableType(propertyType))
             {
@@ -362,7 +366,12 @@ namespace Azure.Search.Documents.Indexes
             public static bool TryGet(Type type, IMemberNameConverter nameProvider, out ObjectInfo info)
             {
                 // Close approximation to Newtonsoft.Json.Serialization.DefaultContractResolver that was used in Microsoft.Azure.Search.
-                if (!type.IsPrimitive && !type.IsEnum && !s_unsupportedTypes.Contains(type) && !s_primitiveTypeMap.ContainsKey(type) && !typeof(IEnumerable).IsAssignableFrom(type))
+                if (!type.IsPrimitive &&
+                    !type.IsEnum &&
+                    !s_unsupportedTypes.Contains(type) &&
+                    !s_primitiveTypeMap.ContainsKey(type) &&
+                    !GeometryFactory.CanCreate(type) &&
+                    !typeof(IEnumerable).IsAssignableFrom(type))
                 {
                     const BindingFlags bindingFlags = BindingFlags.Instance | BindingFlags.Public | BindingFlags.NonPublic;